caching transformUserLessFilesToCss

+++ classes/ExtAutomator.php
@@ -6,22 +6,19 @@ class ExtAutomator extends \Automator
 {
 	public function purgeLessCache()
 	{
-		// Purge the folder
-		$objFolder = new \Folder(BOOTSTRAPLESSCUSTOMDIR);
-		$objFolder->purge();
-
-		// Restore the index.html file
-		$objFile = new \File('templates/index.html', true);
-		$objFile->copyTo(BOOTSTRAPLESSCUSTOMDIR . 'index.html');
-
-		// Recreate the less files
-		$this->import('ExtCss');
-		$this->ExtCss->updateExtCss();
-
-		// Recreate css files
-		$this->import('StyleSheets');
-		$this->StyleSheets->updateStylesheets();
-
+		if(!is_array($GLOBALS['TL_PURGE']['folders']['less']['affected'])) return false;
+		
+		foreach($GLOBALS['TL_PURGE']['folders']['less']['affected'] as $folder)
+		{
+			// Purge folder
+			$objFolder = new \Folder($folder);
+			$objFolder->purge();
+			
+			// Restore the index.html file
+			$objFile = new \File('templates/index.html', true);
+			$objFile->copyTo($folder . 'index.html');
+		}
+		
 		// Also empty the page cache so there are no links to deleted scripts
 		$this->purgePageCache();
